Grouped pg PongoClientOptions
