+++ src/packages/pongo/src/storage/postgresql/pg/index.ts
@@ -1,0 +1,35 @@
+import type { NodePostgresConnection } from '@event-driven-io/dumbo/pg';
+import pg from 'pg';
+
+export type NodePostgresPongoClientOptions =
+  | PooledPongoClientOptions
+  | NotPooledPongoOptions;
+
+export type PooledPongoClientOptions =
+  | {
+      pool: pg.Pool;
+    }
+  | {
+      pooled: true;
+    }
+  | {
+      pool: pg.Pool;
+      pooled: true;
+    }
+  | object;
+
+export type NotPooledPongoOptions =
+  | {
+      client: pg.Client;
+    }
+  | {
+      pooled: false;
+    }
+  | {
+      client: pg.Client;
+      pooled: false;
+    }
+  | {
+      connection: NodePostgresConnection;
+      pooled?: false;
+    };
